refactor: made more universal version
