+++ src/js/changeLanguage.js
@@ -1,10 +1,9 @@
 import getLocalLanguage from './getLocalLanguage';
 
 function changeLanguage() {
-    const language = getLocalLanguage();
-    if (language === 'en') localStorage.setItem('language', 'ru');
-    else localStorage.setItem('language', 'en');
-    document.querySelectorAll('.keyboard__button').forEach((button) => [...button.children].forEach((item) => item.classList.toggle('current')));
+  const language = getLocalLanguage();
+  if (language === 'en') localStorage.setItem('language', 'ru');
+  else localStorage.setItem('language', 'en');
+  document.querySelectorAll('.keyboard__button').forEach((button) => [...button.children].forEach((item) => item.classList.toggle('current')));
 }
-
 export default changeLanguage;
