--- src/js/generalFunctions.js
@@ -1,7 +1,0 @@
-function generateDomElement(currentNode, element, ...classes) {
-    currentNode = document.createElement(element);
-    currentNode.classList.add(...classes);
-    return currentNode;
-}
-
-export default generateDomElement;
