+++ src/js/keyboardView.js
@@ -1,78 +1,78 @@
-import generateDomElement from './generalFunctions';
+import generateDomElement from './helpers';
 import buttonsLayout from './buttonsLayout';
 import getLocalLanguage from './getLocalLanguage';
 
 function createKeyboardView() {
-    const wrapper = generateDomElement(wrapper, 'div', 'wrapper');
-    document.body.append(wrapper);
+  const wrapper = generateDomElement('div', ['wrapper']);
+  document.body.append(wrapper);
 
-    const textarea = generateDomElement(textarea, 'textarea', 'textarea');
-    textarea.setAttribute('autofocus', 'autofocus');
-    textarea.setAttribute('cols', 100);
-    textarea.setAttribute('rows', 15);
-
-
-    const keyboardContainer = generateDomElement(keyboardContainer, 'div', 'keyboard-container');
+  const textarea = generateDomElement('textarea', ['textarea']);
+  textarea.setAttribute('autofocus', 'autofocus');
+  textarea.setAttribute('cols', 100);
+  textarea.setAttribute('rows', 15);
 
 
-    buttonsLayout.forEach((row) => {
-        row.forEach((element) => {
-            const button = generateDomElement(button, 'button');
-            const [code, ruLittle, ruBig, enLittle, enBig] = element;
+  const keyboardContainer = generateDomElement('div', ['keyboard-container']);
 
-            button.classList.add('keyboard__button', code);
-            switch (element[0]) {
-                case 'Space':
-                    button.classList.add('keyboard__button_the-widest');
-                    break;
-                case 'Backspace':
-                case 'CapsLock':
-                case 'ShiftLeft':
-                case 'Enter':
-                    button.classList.add('keyboard__button_wider');
-                    break;
-                case 'ShiftRight':
-                    button.classList.add('keyboard__key-base-width');
-                    break;
 
-                default:
-                    {
-                        break;
-                    }
-            }
+  buttonsLayout.forEach((row) => {
+    row.forEach((element) => {
+      const button = generateDomElement('button');
+      const [code, ruLittle, ruBig, enLittle, enBig] = element;
 
-            keyboardContainer.append(button);
+      button.classList.add('keyboard__button', code);
+      switch (element[0]) {
+        case 'Space':
+          button.classList.add('keyboard__button_the-widest');
+          break;
+        case 'Backspace':
+        case 'CapsLock':
+        case 'ShiftLeft':
+        case 'Enter':
+          button.classList.add('keyboard__button_wider');
+          break;
+        case 'ShiftRight':
+          button.classList.add('keyboard__key-base-width');
+          break;
 
-            const ruLang = generateDomElement(ruLang, 'span', 'ru');
-            const enLang = generateDomElement(ruLang, 'span', 'en');
+        default:
+        {
+          break;
+        }
+      }
 
-            const language = getLocalLanguage();
-            if (language === 'en') enLang.classList.add('current');
-            else ruLang.classList.add('current');
+      keyboardContainer.append(button);
 
-            button.append(ruLang, enLang);
+      const ruLang = generateDomElement('span', ['ru']);
+      const enLang = generateDomElement('span', ['en']);
 
-            const ruLowerCase = generateDomElement(ruLang, 'span', 'small', 'on');
-            ruLowerCase.textContent = ruLittle;
+      const language = getLocalLanguage();
+      if (language === 'en') enLang.classList.add('current');
+      else ruLang.classList.add('current');
 
-            const ruUpperCase = generateDomElement(ruLang, 'span', 'big');
-            ruUpperCase.textContent = ruBig;
-            ruLang.append(ruLowerCase, ruUpperCase);
+      button.append(ruLang, enLang);
 
-            const enLowerCase = generateDomElement(ruLang, 'span', 'small', 'on');
-            enLowerCase.textContent = enLittle;
+      const ruLowerCase = generateDomElement('span', ['small', 'on']);
+      ruLowerCase.textContent = ruLittle;
 
-            const enUpperCase = generateDomElement(ruLang, 'span', 'big');
-            enUpperCase.textContent = enBig;
-            enLang.append(enLowerCase, enUpperCase);
-        });
-        keyboardContainer.append(document.createElement('br'));
+      const ruUpperCase = generateDomElement('span', ['big']);
+      ruUpperCase.textContent = ruBig;
+      ruLang.append(ruLowerCase, ruUpperCase);
+
+      const enLowerCase = generateDomElement('span', ['small', 'on']);
+      enLowerCase.textContent = enLittle;
+
+      const enUpperCase = generateDomElement('span', ['big']);
+      enUpperCase.textContent = enBig;
+      enLang.append(enLowerCase, enUpperCase);
     });
+    keyboardContainer.append(document.createElement('br'));
+  });
 
-    const textAboutChangingLanguage = generateDomElement(textAboutChangingLanguage, 'span', 'about-language');
-    textAboutChangingLanguage.innerText = 'For changing language use \'ShiftLeft\' +  \'AltLeft\' button combination :)';
+  const textAboutChangingLanguage = generateDomElement('span', ['about-language']);
+  textAboutChangingLanguage.innerText = 'For changing language use \'ShiftLeft\' +  \'AltLeft\' button combination :)';
 
-    wrapper.append(textarea, keyboardContainer, textAboutChangingLanguage);
+  wrapper.append(textarea, keyboardContainer, textAboutChangingLanguage);
 }
 
 export default createKeyboardView;
